software/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
RPi.GPIO==0.7.1
Adafruit_GPIO==1.0.4
Adafruit_SSD1306==1.6.2
Adafruit_CharLCD==1.0.0
hx711==1.1.1
schedule==1.2.0
psutil==5.9.5
requests==2.31.0
python-dateutil==2.8.2
Pillow==10.0.1
numpy==1.24.3
matplotlib==3.7.2
pandas==2.0.3 